refactor: rename position files for better organization
