--- packages/auto-consensus/src/position.ts
@@ -1,5 +1,0 @@
-// file: src/position.ts
-
-export * from './position/index'
-export * from './position/price'
-export * from './position/utils'
